@@ -1,39 +1,45 @@
 #pragma once
 #include "Color.h"
-#include <vector>
 #include <iostream>
+#include <vector>
+enum struct ColorSchemeType {
+    Monochrome,
+    Complementary,
+    Triadic,
+    Analogous
+};
 
-struct ColorScheme{
+struct ColorScheme {
 protected:
     std::vector<double> hues;
 
 public:
     ColorScheme(std::vector<double> hues);
-    const std::vector<double>& GetHues() const { return hues; }
+    const std::vector<double> &GetHues() const { return hues; }
     int GetNumberOfHues() const { return hues.size(); }
 };
 
-struct MonochromeColorScheme : public ColorScheme{
+struct MonochromeColorScheme : public ColorScheme {
 public:
-    MonochromeColorScheme(double hue): ColorScheme({hue}){}
+    MonochromeColorScheme(double hue) : ColorScheme({hue}) {}
 };
 
-struct ComplementaryColorScheme : public ColorScheme{
+struct ComplementaryColorScheme : public ColorScheme {
 public:
-    ComplementaryColorScheme(double hue): ColorScheme({hue, hue + 180}){}
+    ComplementaryColorScheme(double hue) : ColorScheme({hue, hue + 180}) {}
 };
 
-struct AnalogousColorScheme : public ColorScheme{
+struct AnalogousColorScheme : public ColorScheme {
 public:
-    AnalogousColorScheme(double hue, double colorDistance, unsigned int numberOfHues): ColorScheme({hue}){
-        for(unsigned int i = 1; i < (numberOfHues - 1) / 2; i++){
+    AnalogousColorScheme(double hue, double colorDistance, unsigned int numberOfHues) : ColorScheme({hue}) {
+        for (unsigned int i = 1; i < (numberOfHues - 1) / 2; i++) {
             this->hues.push_back(hue + i * colorDistance);
             this->hues.push_back(hue - i * colorDistance);
         }
     }
 };
 
-struct TriadicColorScheme : public ColorScheme{
+struct TriadicColorScheme : public ColorScheme {
 public:
-    TriadicColorScheme(double hue): ColorScheme({hue, hue + 120, hue + 240}){}
+    TriadicColorScheme(double hue) : ColorScheme({hue, hue + 120, hue + 240}) {}
 };
